requirements.txt
annotated-types==0.7.0
anyio==4.12.1
appopener==1.7
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.15
cryptography==46.0.4
customtkinter==5.2.2
darkdetect==0.8.0
distro==1.9.0
duckduckgo_search==8.1.1
Flask==3.1.2
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.63.0
googleapis-common-protos==1.72.0
googlesearch-python==1.3.0
groq==1.0.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
keyboard==0.13.5
lxml==6.0.2
MarkupSafe==3.0.3
MouseInfo==0.1.3
openai==2.18.0
packaging==26.0
pillow==12.1.1
primp==0.15.0
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20251023
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyparsing==3.3.2
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pyttsx3==2.99
pytweening==1.2.0
pywhatkit==5.4
pywin32==311
pywinauto==0.6.9
regex==2026.1.15
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SpeechRecognition==3.14.5
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uiautomation==2.0.29
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
Werkzeug==3.1.5
wikipedia==1.4.0
word2number==1.1
